+++ library/hls/src/main/java/com/google/android/exoplayer2/source/hls/HlsMediaSource.java
@@ -23,8 +23,11 @@ import com.google.android.exoplayer2.ExoPlayerLibraryInfo;
 import com.google.android.exoplayer2.extractor.Extractor;
 import com.google.android.exoplayer2.source.AdaptiveMediaSourceEventListener;
 import com.google.android.exoplayer2.source.AdaptiveMediaSourceEventListener.EventDispatcher;
+import com.google.android.exoplayer2.source.CompositeSequenceableLoaderFactory;
+import com.google.android.exoplayer2.source.DefaultCompositeSequenceableLoaderFactory;
 import com.google.android.exoplayer2.source.MediaPeriod;
 import com.google.android.exoplayer2.source.MediaSource;
+import com.google.android.exoplayer2.source.SequenceableLoader;
 import com.google.android.exoplayer2.source.SinglePeriodTimeline;
 import com.google.android.exoplayer2.source.hls.playlist.HlsMediaPlaylist;
 import com.google.android.exoplayer2.source.hls.playlist.HlsPlaylist;
@@ -59,6 +62,8 @@ public final class HlsMediaSource implements MediaSource,
     private ParsingLoadable.Parser<HlsPlaylist> playlistParser;
     private AdaptiveMediaSourceEventListener eventListener;
     private Handler eventHandler;
+    private CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory;
+
     private int minLoadableRetryCount;
     private boolean isBuildCalled;
 
@@ -151,6 +156,22 @@ public final class HlsMediaSource implements MediaSource,
     }
 
     /**
+     * Sets the factory to create composite {@link SequenceableLoader}s for when this media source
+     * loads data from multiple streams (video, audio etc...). The default is an instance of
+     * {@link DefaultCompositeSequenceableLoaderFactory}.
+     *
+     * @param compositeSequenceableLoaderFactory A factory to create composite
+     *     {@link SequenceableLoader}s for when this media source loads data from multiple streams
+     *     (video, audio etc...).
+     * @return This builder.
+     */
+    public Builder setCompositeSequenceableLoaderFactory(
+        CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory) {
+      this.compositeSequenceableLoaderFactory = compositeSequenceableLoaderFactory;
+      return this;
+    }
+
+    /**
      * Builds a new {@link HlsMediaSource} using the current parameters.
      * <p>
      * After this call, the builder should not be re-used.
@@ -167,8 +188,12 @@ public final class HlsMediaSource implements MediaSource,
       if (playlistParser == null) {
         playlistParser = new HlsPlaylistParser();
       }
+      if (compositeSequenceableLoaderFactory == null) {
+        compositeSequenceableLoaderFactory = new DefaultCompositeSequenceableLoaderFactory();
+      }
       return new HlsMediaSource(manifestUri, hlsDataSourceFactory, extractorFactory,
-          minLoadableRetryCount, eventHandler, eventListener, playlistParser);
+          compositeSequenceableLoaderFactory, minLoadableRetryCount, eventHandler, eventListener,
+          playlistParser);
     }
 
   }
@@ -181,6 +206,7 @@ public final class HlsMediaSource implements MediaSource,
   private final HlsExtractorFactory extractorFactory;
   private final Uri manifestUri;
   private final HlsDataSourceFactory dataSourceFactory;
+  private final CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory;
   private final int minLoadableRetryCount;
   private final EventDispatcher eventDispatcher;
   private final ParsingLoadable.Parser<HlsPlaylist> playlistParser;
@@ -242,11 +268,23 @@ public final class HlsMediaSource implements MediaSource,
       HlsExtractorFactory extractorFactory, int minLoadableRetryCount, Handler eventHandler,
       AdaptiveMediaSourceEventListener eventListener,
       ParsingLoadable.Parser<HlsPlaylist> playlistParser) {
+    this(manifestUri, dataSourceFactory, extractorFactory,
+        new DefaultCompositeSequenceableLoaderFactory(), minLoadableRetryCount, eventHandler,
+        eventListener, playlistParser);
+  }
+
+  private HlsMediaSource(Uri manifestUri, HlsDataSourceFactory dataSourceFactory,
+      HlsExtractorFactory extractorFactory,
+      CompositeSequenceableLoaderFactory compositeSequenceableLoaderFactory,
+      int minLoadableRetryCount, Handler eventHandler,
+      AdaptiveMediaSourceEventListener eventListener,
+      ParsingLoadable.Parser<HlsPlaylist> playlistParser) {
     this.manifestUri = manifestUri;
     this.dataSourceFactory = dataSourceFactory;
     this.extractorFactory = extractorFactory;
     this.minLoadableRetryCount = minLoadableRetryCount;
     this.playlistParser = playlistParser;
+    this.compositeSequenceableLoaderFactory = compositeSequenceableLoaderFactory;
     eventDispatcher = new EventDispatcher(eventHandler, eventListener);
   }
 
@@ -268,7 +306,7 @@ public final class HlsMediaSource implements MediaSource,
   public MediaPeriod createPeriod(MediaPeriodId id, Allocator allocator) {
     Assertions.checkArgument(id.periodIndex == 0);
     return new HlsMediaPeriod(extractorFactory, playlistTracker, dataSourceFactory,
-        minLoadableRetryCount, eventDispatcher, allocator);
+        minLoadableRetryCount, eventDispatcher, allocator, compositeSequenceableLoaderFactory);
   }
 
   @Override